requirements.txt
flask
flask-cors
gunicorn
resend
markdown
python-dotenv
requests
